Bunch of shit that I can't keep track of.
Add new LipSync modes & Smoothing to GUI.
Bunch of cleanup.
Added logging for some simple things that came to mind.
Fixed applyRootMotion being enabled.

Start of AnimatorManager. It's a bit messy but works for now while I get the motivation to finish the in-scene player controller feature. Was actually written for a different project, so kinda shoved it in with little changes.